# Conflicts:
#	src/Application/Adapter/JobQueueInterface.php
#	src/Application/Command/Sheet/Batch.php
#	src/Application/Command/Sheet/BatchAssignHandler.php
#	src/Application/Command/Sheet/BatchDraft.php
#	src/Application/Command/Sheet/BatchHandler.php
#	src/Infrastructure/Bundle/InfrastructureBundle/Adapter/JobQueueAdapter.php
#	src/Infrastructure/Bundle/InfrastructureBundle/Resources/config/services.yml
#	src/Ui/Bundle/AdminBundle/Controller/SheetController.php
#	src/Ui/Bundle/EventBundle/Resources/config/services.yml

+++ src/Application/Command/Unavailability/RemoveHandler.php
@@ -10,7 +10,11 @@
 
 namespace Proximum\Vimeet\Application\Command\Unavailability;
 
+use Proximum\Vimeet\Application\Exception\Unavailability\CanNotDeleteUnavailabilityException;
+use Proximum\Vimeet\Application\Event\Events;
+use Proximum\Vimeet\Application\Event\Unavailability\RemoveUnavailabilityEvent;
 use Proximum\Vimeet\Domain\Repository\UnavailabilityRepositoryInterface;
+use Proximum\Vimeet\Infrastructure\Adapter\DelayedEventDispatcher;
 
 class RemoveHandler
 {
@@ -19,21 +23,39 @@ class RemoveHandler
      */
     private $unavailabilityRepository;
 
+    /** @var DelayedEventDispatcher */
+    private $eventDispatcher;
+
     /**
      * RemoveHandler constructor.
      *
      * @param UnavailabilityRepositoryInterface $unavailabilityRepository
+     * @param DelayedEventDispatcher            $eventDispatcher
      */
-    public function __construct(UnavailabilityRepositoryInterface $unavailabilityRepository)
-    {
+    public function __construct(
+        UnavailabilityRepositoryInterface $unavailabilityRepository,
+        DelayedEventDispatcher $eventDispatcher
+    ) {
         $this->unavailabilityRepository = $unavailabilityRepository;
+        $this->eventDispatcher          = $eventDispatcher;
     }
 
     /**
      * @param Remove $remove
+     *
+     * @throws CanNotDeleteUnavailabilityException
      */
     public function handle(Remove $remove)
     {
+        if (!$remove->unavailability->getParticipant()->getSheet()->attend()) {
+            throw new CanNotDeleteUnavailabilityException('The sheet does not attend the event, therefore the unavailability can not be remove');
+        }
+
         $this->unavailabilityRepository->remove($remove->unavailability);
+
+        $this->eventDispatcher->dispatch(
+            Events::UNAVAILABILITY_REMOVED,
+            new RemoveUnavailabilityEvent($remove->unavailability->getParticipant())
+        );
     }
 }
